requirements.txt
opencv-contrib-python
pytorch
pushbullet
pigpio